PMA-127 - ModulePropertyPolicy

+++ app/Http/Controllers/ModulePropertyController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ModuleProperty\UpdateRequest;
 use App\Http\Resources\ModulePropertyResource;
 use App\Http\Resources\ModulePropertyResourceCollection;
 use App\Repositories\Contracts\ModulePropertyRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class ModulePropertyController extends Controller
 {
@@ -32,9 +33,12 @@ class ModulePropertyController extends Controller
      *
      * @param IndexRequest $request
      * @return ModulePropertyResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', ModuleProperty::clearBootedModels());
+
         $moduleProperties = $this->modulePropertyRepository->findBy($request->all());
 
         return new ModulePropertyResourceCollection($moduleProperties);
@@ -45,9 +49,12 @@ class ModulePropertyController extends Controller
      *
      * @param StoreRequest  $request
      * @return ModulePropertyResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', ModuleProperty::clearBootedModels());
+
         $moduleProperty = $this->modulePropertyRepository->setModuleProperty($request->all());
 
         return new ModulePropertyResource($moduleProperty);
@@ -58,9 +65,12 @@ class ModulePropertyController extends Controller
      *
      * @param ModuleProperty $moduleProperty
      * @return ModulePropertyResource
+     * @throws AuthorizationException
      */
     public function show(ModuleProperty $moduleProperty)
     {
+        $this->authorize('show', $moduleProperty);
+
         return new ModulePropertyResource($moduleProperty);
     }
 
@@ -70,9 +80,12 @@ class ModulePropertyController extends Controller
      * @param UpdateRequest $request
      * @param ModuleProperty $moduleProperty
      * @return ModulePropertyResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ModuleProperty $moduleProperty)
     {
+        $this->authorize('update', $moduleProperty);
+
         $moduleProperty = $this->modulePropertyRepository->setModuleProperty($moduleProperty, $request->all());
 
         return new ModulePropertyResource($moduleProperty);
@@ -83,9 +96,12 @@ class ModulePropertyController extends Controller
      *
      * @param ModuleProperty $moduleProperty
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(ModuleProperty $moduleProperty)
     {
+        $this->authorize('destroy', $moduleProperty);
+
         $this->modulePropertyRepository->delete($moduleProperty);
 
         return response()->json(null, 204);
